% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking.R
\name{check_label}
\alias{check_label}
\title{Check label type}
\usage{
check_label(main, set)
}
\arguments{
\item{main}{main data object}

\item{set}{config object}
}
\value{
set
}
\description{
This function makes sure the set$main$labeliscategory
parameter is correctly set, according to the data type
of the label column.
}
