% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking.R
\name{check_missing}
\alias{check_missing}
\title{Check for Missing Data.}
\usage{
check_missing(main, set, prep)
}
\arguments{
\item{main}{main data object}

\item{set}{config object}

\item{prep}{summary object}
}
\description{
This function can be used to halt the workflow, if the data contains
missing values.
}
