% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R, R/wrap_up.R
\name{clear_model_results}
\alias{clear_model_results}
\title{Clearing result tables}
\usage{
clear_model_results(result_path)

clear_model_results(result_path)

clear_model_results(result_path)
}
\arguments{
\item{result_path}{}
}
\description{
This function clears result tables
by removing all rows, except for the
column names. As input, give the path
where the \code{result} folder is located
(e.g., \code{"project_path/output_model"}).

This function clears result tables
by removing all rows, except for the
column names. As input, give the path
where the \code{result} folder is located.

This function clears information from result tables
}
