% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{cluster_dataprep}
\alias{cluster_dataprep}
\title{Wrapper Function for preparing data to apply clustering methods.}
\usage{
cluster_dataprep(X, set)
}
\arguments{
\item{X}{data object}

\item{set}{config object}
}
\value{
list of data frames and distance matrix.
}
\description{
Wrapper Function for preparing data to apply clustering methods.
}
