% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R, R/wrap_up.R
\name{collect_results}
\alias{collect_results}
\title{Collecting results back to R from local files}
\usage{
collect_results(
  result_path,
  executionid,
  tables = c("execution", "accuracy", "column_importance", "coefficients", "metadata",
    "validation")
)

collect_results(
  result_path,
  executionid,
  tables = c("execution", "accuracy", "column_importance", "coefficients", "metadata",
    "validation")
)
}
\arguments{
\item{result_path}{path to \code{result} directory}

\item{executionid}{for which execution to get the results}

\item{tables}{which result tables to return?}
}
\description{
Collecting results back to R from local files

Collect desired results from local files.
}
