% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive_service.r
\name{create_df}
\alias{create_df}
\title{Parse data.frame from input parameters}
\usage{
create_df(param, param2, param3)
}
\arguments{
\item{param}{API parameter of data values}

\item{param2}{API parameter of column names}

\item{param3}{API parameter of data types}
}
\value{
data.frame object
}
\description{
Parse data.frame from input parameters
}
