% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2o_train.R
\name{create_models}
\alias{create_models}
\title{Wrapper Function for creating models & outputs}
\usage{
create_models(set, main, prep, odbc)
}
\arguments{
\item{set}{config object}

\item{main}{main data object}

\item{prep}{summary object}

\item{odbc}{ODBC connection object (only needed when using DB connection)}
}
\value{
(1) Model fit measures, (2) Featurwe importances (for ML models),
(3) Model coefficients (GLM models),
(4) Model predictions for VALIDATION data,
(5) Table for model application, (6) Data factor levels, and
(7) Model parameters
}
\description{
Wrapper Function for creating models & outputs
}
