% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2o_predict.R
\name{create_predictions}
\alias{create_predictions}
\title{Get model predictions.}
\usage{
create_predictions(
  df,
  set,
  prep,
  apply_models,
  odbc,
  odbc_pred,
  useMOJO = TRUE
)
}
\arguments{
\item{df}{main data object}

\item{set}{config file}

\item{prep}{summary object}

\item{apply_models}{row specifying applied model}

\item{odbc}{ODBC connection config}

\item{odbc_pred}{ODBC connection config for model application}

\item{useMOJO}{whether to use MOJO-models or H2O-formated
models (default = TRUE. At present, storing H2O-models is
disabled in the \code{export_model_output} function).}
}
\value{
model predictions
}
\description{
Get model predictions.
}
