% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{create_split}
\alias{create_split}
\title{Create data partitioning key}
\usage{
create_split(main, set)
}
\arguments{
\item{main}{main data object}

\item{set}{config object}
}
\value{
main data object with data split column added
}
\description{
If the label is categorical, a stratified data split
(based on the label variable) is created using
the createDataPartition function in the caret-package,
based on split proportions defined in configuration. Othervise,
a random split is made. A new column is added to the type-set
data, which is returned.
}
