% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_connections.R
\name{csv_preparations}
\alias{csv_preparations}
\title{Generate execution tables}
\usage{
csv_preparations(set, main)
}
\arguments{
\item{set}{config object}

\item{main}{main data object}
}
\value{
list with runid,execution_row, summary_table, and columns
}
\description{
Function for getting execution tables, when using
local file connection.
}
