% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{data_read}
\alias{data_read}
\title{Wrapper Function for Reading Data from Source.}
\usage{
data_read(set, odbc)
}
\arguments{
\item{set}{config object}

\item{odbc}{ODBC connection object}
}
\value{
main data object
}
\description{
Wrapper Function for Reading Data from Source.
}
