% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_connections.R
\name{db_preparations}
\alias{db_preparations}
\title{Establishing ODBC connection}
\usage{
db_preparations(set, variable_types, df, odbc_metadata)
}
\arguments{
\item{set}{config object}

\item{variable_types}{variable types table}

\item{df}{data object}

\item{odbc_metadata}{}
}
\value{
A list of odbc read and write connections.
}
\description{
Establishing ODBC connection
}
\examples{
db_preparations(set, variable_types, df, odbc_metadata)

}
