% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_up.R
\name{delete_project}
\alias{delete_project}
\title{Delete a project from given path.}
\usage{
delete_project(project_path)
}
\arguments{
\item{path}{system path from where to delete a project}
}
\description{
Delete a project from given path.
}
