% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{do_transforms}
\alias{do_transforms}
\title{Wrapper Function for Making Data Transforms.}
\usage{
do_transforms(
  main,
  set,
  prep,
  selection = c("delete_equal", "clean_special", "classify_NA", "delete_constant")
)
}
\arguments{
\item{main}{main data object}

\item{set}{config object}

\item{prep}{summary object}

\item{selection}{which transformations to apply?}
}
\value{
main data object with transformed data.
}
\description{
allowed transformations \code{c("delete_equal", "clean_special", 
"classify_NA", "delete_constant")}. In addition, a column
defining a data split will also be added, if missing from
the data.
}
\examples{
do_transforms(main, set, prep,
              selection = c("delete_equal",
                            "clean_special",
                            "classify_NA",
                            "delete_constant"))

}
