% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{entropy_recategorization}
\alias{entropy_recategorization}
\title{Wrapper Function for calling entropy transformation.}
\usage{
entropy_recategorization(main, set, prep)
}
\arguments{
\item{main}{main data object}

\item{set}{config object}

\item{prep}{summary object}
}
\value{
main object with transformed data.
}
\description{
Wrapper Function for calling entropy transformation.
}
