% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{export_clust_output}
\alias{export_clust_output}
\title{Function for saving models results to output.}
\usage{
export_clust_output(output, set, prep, odbc)
}
\arguments{
\item{output}{clustering results}

\item{set}{config object}

\item{prep}{config object}

\item{odbc}{db connection}
}
\description{
Function for saving models results to output.
}
