% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2o_utilities.R
\name{export_model_output}
\alias{export_model_output}
\title{Function for writing model outputs}
\usage{
export_model_output(models, output, set, prep, odbc)
}
\arguments{
\item{models}{a list of trained models}

\item{output}{a list of generated model outputs}

\item{set}{config object}

\item{prep}{summary object}

\item{odbc}{ODBC connection file (only needed when using DB connection)}
}
\value{
(1) Model fit measures, Featurwe importances (for ML models),
Model coefficients (GLM models),
Model predictions for VALIDATION data, and
Table for model application are written to tables (either locl or DB)
(2) Data factor levels and Model parameters are written to local .Rds files
(3) Models are exported as MOJO objects.
}
\description{
Function for writing model outputs
}
