% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2o_utilities.R
\name{fix_metric}
\alias{fix_metric}
\title{Function for making sure evaluation metric is compatible with H2O.}
\usage{
fix_metric(set)
}
\arguments{
\item{set}{config object}
}
\value{
fixed metric
}
\description{
Function for making sure evaluation metric is compatible with H2O.
}
