% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2o_optimize.R
\name{get_best_params}
\alias{get_best_params}
\title{Return best parameters for given models}
\usage{
get_best_params(grid, model_ids, metric)
}
\arguments{
\item{grid}{H2O model grid}

\item{model_ids}{model-ids}

\item{metric}{evaluation metric}
}
\value{
best parameters
}
\description{
Return best parameters for given models
}
