% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{get_cluster_output}
\alias{get_cluster_output}
\title{Function for saving models results to output.}
\usage{
get_cluster_output(set, runid, data)
}
\arguments{
\item{set}{config object}

\item{data}{data object}
}
\value{
list of algorithms results
}
\description{
Function for saving models results to output.
}
