% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling_preparations.R
\name{get_levels}
\alias{get_levels}
\title{Function for extracting and factor levels}
\usage{
get_levels(df)
}
\arguments{
\item{df}{data.frame}
}
\value{
list of factor levels per feature
}
\description{
Function for extracting and factor levels
}
