% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2o_utilities.R
\name{get_model_output}
\alias{get_model_output}
\title{Function for generating model outputs}
\usage{
get_model_output(models, model_names, set, runid, h2o_data)
}
\arguments{
\item{models}{either a single model object or a list of models}

\item{set}{config object}

\item{runid}{run ID}

\item{h2o_data}{a list of H2OFrames}
}
\value{
data.frame of model performance, data.frame of feature importances/model coefficients,
model predictions for VALIDATION set, table of model ID's and application tag, and
a list of factor levels for categorical predictors.
}
\description{
Function for generating model outputs
}
