% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{handling_trycatch}
\alias{handling_trycatch}
\title{Catch errors, warnings, execution time}
\usage{
handling_trycatch(expr)
}
\arguments{
\item{expr}{function call}
}
\description{
This module is used in front of all functions and
commands
}
\examples{
handling_trycatch(data_read(set, odbc))

}
