% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{init_aijack}
\alias{init_aijack}
\title{Creates a project template to given
path}
\usage{
init_aijack(path)
}
\arguments{
\item{path}{system path where to create a project}
}
\description{
The function creates a new directory (if it
does not exist) and within that directory, four sub-
directories: \code{output_model}, \code{output_apply},
\code{source_model}, and \code{source_apply}. Directories
\code{output_model} and \code{output_apply} will be further
populated by \code{.scv} files that contain output tables.
In addition, a directory with name \code{control} is created,
which is populated with .R scripts for controlling execution.
To delete a project, call \code{delete_project(path)}
function.
}
