% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive_service.r
\name{insert_db_plumber}
\alias{insert_db_plumber}
\title{Insert predictions to DB table}
\usage{
insert_db_plumber(channel, table_name, prediction, apply_models)
}
\arguments{
\item{channel}{odbc connection}

\item{table_name}{table name in DB}

\item{prediction}{data.frame with model prediction}

\item{apply_models}{selected row from models-table}
}
\value{
Writes predictions to DB table.
}
\description{
Insert predictions to DB table
}
