% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2o_predict.R
\name{list_models}
\alias{list_models}
\title{List models}
\usage{
list_models(model_path)
}
\description{
This function lists available models
that can be found in the \code{models}
results table, located in the
\code{model_path} directory, which
typically is \code{"output_model/results"}
}
