% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{logging_control}
\alias{logging_control}
\title{Handling excution logs}
\usage{
logging_control(set)
}
\arguments{
\item{set}{config object}
}
\description{
Depending on settings, either prints logs to
console or writes them files.
}
