% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2o_utilities.R
\name{make_h2o_data}
\alias{make_h2o_data}
\title{Convert data-splits to H2O frames}
\usage{
make_h2o_data(df, set)
}
\arguments{
\item{df}{main data object, from which the "splitted"-object is used}

\item{set}{config object}
}
\value{
a list containing training ("train"),
testing ("test"), and validation ("val")
sets, as well as a combined train-test
set ("full_train").
}
\description{
Convert data-splits to H2O frames
}
