% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2o_predict.R
\name{make_predictions}
\alias{make_predictions}
\title{Wrapper Function for creating model predictions}
\usage{
make_predictions(set, main, prep, odbc)
}
\arguments{
\item{set}{config object}

\item{main}{main data object}

\item{prep}{summary object}

\item{odbc}{ODBC connection object (only needed when using DB connection)}
}
\description{
This fucntion calls the \code{create_predictions()} function.
}
