% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling_preparations.R
\name{make_tables}
\alias{make_tables}
\title{Function for generating tables for execution,
metadata and data columns}
\usage{
make_tables(runid, df, types, query)
}
\arguments{
\item{runid}{execution ID}

\item{df}{data.frame}

\item{types}{variable types}

\item{query}{data load call}
}
\value{
list with execution_row, summary_table, and columns
}
\description{
Function for generating tables for execution,
metadata and data columns
}
