% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_connections.R
\name{odbc_connections}
\alias{odbc_connections}
\title{Establishing ODBC connection}
\usage{
odbc_connections(set)
}
\arguments{
\item{set}{config object}
}
\value{
A list of odbc read and write connections.
}
\description{
Establishing ODBC connection
}
