% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_connections.R
\name{open_connections}
\alias{open_connections}
\title{Wrapper Function for Establishing Data Source
Connections.}
\usage{
open_connections(set)
}
\arguments{
\item{set}{= config object}
}
\value{
ODBC connection object (NULL if using file
connection)
}
\description{
Wrapper Function for Establishing Data Source
Connections.
}
