% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{optimize_clustering}
\alias{optimize_clustering}
\title{Function for applying clustering algorithms.}
\usage{
optimize_clustering(data, set)
}
\arguments{
\item{data}{data object}

\item{set}{config object}
}
\value{
list of algorithms results.
}
\description{
Function for applying clustering algorithms.
}
