% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2o_optimize.R
\name{optimize_deeplearning_parameters}
\alias{optimize_deeplearning_parameters}
\title{Optimizing DeepLearning parameters}
\usage{
optimize_deeplearning_parameters(
  h2o_data,
  set,
  metric,
  predictors,
  targets,
  runid
)
}
\arguments{
\item{h2o_data}{a list of H2OFrames}

\item{set}{config object}

\item{metric}{evaluation metric}

\item{predictors}{vector of explanatory variables}

\item{targets}{label variable}

\item{runid}{run ID}
}
\value{
best model
}
\description{
Optimizing DeepLearning parameters
}
