% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2o_optimize.R
\name{optimize_drf_parameters}
\alias{optimize_drf_parameters}
\title{Optimizing DRF parameters}
\usage{
optimize_drf_parameters(
  h2o_data,
  set,
  estimator,
  metric,
  predictors,
  targets,
  runid
)
}
\arguments{
\item{h2o_data}{a list of H2OFrames}

\item{set}{config object}

\item{estimator}{either \code{'decisionTree'} or \code{'randomForest'}}

\item{metric}{evaluation metric}

\item{predictors}{vector of explanatory variables}

\item{targets}{label variable}

\item{runid}{run ID}
}
\value{
best model
}
\description{
Optimizing DRF parameters
}
