% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2o_optimize.R
\name{optimize_gbm_parameters}
\alias{optimize_gbm_parameters}
\title{Optimizing GBM parameters.}
\usage{
optimize_gbm_parameters(
  h2o_data,
  set,
  algorithm,
  metric,
  predictors,
  targets,
  runid
)
}
\arguments{
\item{h2o_data}{a list of H2OFrames}

\item{set}{config object}

\item{algorithm}{either \code{'gbm'} or \code{'xgboost'}}

\item{metric}{evaluation metric}

\item{predictors}{vector of explanatory variables}

\item{targets}{label variable}

\item{runid}{run ID}
}
\value{
best model
}
\description{
Optimazation is done in two steps: (1) maximum depth is optimized,
(2) the rest of the parameters are optimized. This procedure makes the
optimization problem more manageable.
}
