% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2o_optimize.R
\name{optimize_glm_parameters}
\alias{optimize_glm_parameters}
\title{Optimizing GLM parameters.}
\usage{
optimize_glm_parameters(h2o_data, set, predictors, targets, runid)
}
\arguments{
\item{h2o_data}{a list of H2OFrames}

\item{set}{config object}

\item{predictors}{vector of explanatory variables}

\item{targets}{label variable}

\item{runid}{run ID}
}
\value{
best model
}
\description{
Optimizing GLM parameters.
}
