% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive_service.r
\name{parse_params}
\alias{parse_params}
\title{Parse input for plumber API}
\usage{
parse_params(file_path, row = 1, set)
}
\arguments{
\item{file_path}{path to file that contains row(s) for prediction}

\item{row}{which row to use (default = 1)?}

\item{set}{plumber config object}
}
\value{
parameter string
}
\description{
Parse input for plumber API
}
