% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive_service.r
\name{plumber_predict}
\alias{plumber_predict}
\title{Get model predictions through plumber API}
\usage{
plumber_predict(df, set, param, param2, param3, odbc)
}
\arguments{
\item{df}{data.frame returned by \code{create_df()}}

\item{set}{plumber config object}

\item{param}{API parameter of data values}

\item{param2}{API parameter of column names}

\item{param3}{API parameter of data types}

\item{odbc}{optional ODBC connection object}
}
\value{
Writes the output to file or to DB, depending on the settings.
}
\description{
Get model predictions through plumber API
}
