% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{prep_results}
\alias{prep_results}
\title{Generate prep-object (summary)}
\usage{
prep_results(set, main, odbc)
}
\arguments{
\item{set}{config object}

\item{main}{main data object}

\item{odbc}{DB connection object}
}
\value{
prep object
}
\description{
Generate prep-object (summary)
}
