% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{print_time}
\alias{print_time}
\title{Function for printing execution time}
\usage{
print_time(start)
}
\arguments{
\item{start}{time of execution start}
}
\description{
Function for printing execution time
}
