% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_csv}
\alias{read_csv}
\title{Read data from CSV files}
\usage{
read_csv(set)
}
\arguments{
\item{set}{config object}
}
\value{
data.frame object.
}
\description{
Uses either \code{read.table} (\code{base}) or
\code{fread} (\code{data.table}), depending on settings. The data is read
from the directory \code{'source_model'/'source_apply'} within the
project folder. If there are several files, the one correct one
is identified by mapping the \code{'model_name_part'}-parameter to the
file names to find a match.
}
