% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_db}
\alias{read_db}
\title{Read data from DB}
\usage{
read_db(query, odbc_read)
}
\arguments{
\item{query}{query statement from config object}

\item{odbc_read}{ODBC-source from config object}
}
\value{
data.frame object.
}
\description{
Uses either \code{read.table} (\code{base}) or
\code{fread} (\code{data.table}), depending on settings.
}
