% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_variabletypes}
\alias{read_variabletypes}
\title{Set variable types in data}
\usage{
read_variabletypes(set, df, odbc_read)
}
\arguments{
\item{set}{config object}

\item{df}{data.frame}

\item{odbc_read}{ODBC-source from config object}
}
\value{
data.frame object.
}
\description{
Set variable types in data
}
