% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_up.R
\name{save_data}
\alias{save_data}
\title{Write main data object to file.}
\usage{
save_data(df, set, prep, get_datetime)
}
\arguments{
\item{df}{data.frame}

\item{set}{config object}

\item{prep}{summary object}

\item{get_datetime}{timestamp}
}
\description{
Write main data object to file.
}
