% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2o_train.R
\name{score_models}
\alias{score_models}
\title{Scoring models}
\usage{
score_models(set, model_ids, metric)
}
\arguments{
\item{set}{config object}

\item{model_ids}{model identifiers}

\item{metric}{evaluation metric}
}
\value{
best model
}
\description{
Scoring models
}
