% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{set_type}
\alias{set_type}
\title{Set data types}
\usage{
set_type(types, column_name, df_i, set)
}
\arguments{
\item{types}{data.frame with variable types}

\item{column_name}{name of variable to be set}

\item{df_i}{column in data.frame to be set}

\item{set}{config object}
}
\value{
data.frame column
}
\description{
Set data types
}
