% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling_preparations.R
\name{split_data}
\alias{split_data}
\title{Function for Data Splitting}
\usage{
split_data(main, set)
}
\arguments{
\item{main}{main data object}

\item{set}{config object}
}
\value{
main object with splitted data.
}
\description{
Split data based on the \code{"test_train_val"}
column in the data table.
}
