% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_stats.R
\name{stat_correlation}
\alias{stat_correlation}
\title{Calculate correlation coefficients.}
\usage{
stat_correlation(df, set, cors)
}
\arguments{
\item{df}{data.frame object}

\item{set}{config object}

\item{cors}{correlation methods to be used (\code{'pearson'},\code{'spearman'}, or \code{'kendall'})}
}
\value{
list of pair-wise correlations per method.
}
\description{
Calculate correlation coefficients.
}
\details{
Automatically handels feature tyopes, such that only
numerical variables are used in operations.
}
\examples{
stat_correlation(df, set, cors = "pearson")

}
