% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2o_train.R
\name{train_autoencoder_model}
\alias{train_autoencoder_model}
\title{Training a anomaly detection model, using autoencoder network}
\usage{
train_autoencoder_model(model_name, h2o_data, set)
}
\arguments{
\item{model_name}{model_id}

\item{h2o_data}{list of H2OFFrames}

\item{set}{config object}
}
\value{
autoencoder model
}
\description{
Training a anomaly detection model, using autoencoder network
}
