% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2o_train.R
\name{train_automl_model}
\alias{train_automl_model}
\title{Trainign H2O models using AutoML}
\usage{
train_automl_model(df, set, runid)
}
\arguments{
\item{df}{list of H2OFrames}

\item{set}{config object}

\item{runid}{run ID}
}
\value{
a list of n best models (n is defines in set) and performance of the best models
}
\description{
Trainign H2O models using AutoML
}
