% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2o_train.R
\name{train_deeplearning_model}
\alias{train_deeplearning_model}
\title{Training DeepLearning model, using optimized parameters}
\usage{
train_deeplearning_model(model_name, h2o_data, best_params)
}
\arguments{
\item{model_name}{model_id}

\item{h2o_data}{list of H2OFFrames}

\item{best_params}{optimised parameters}
}
\value{
DeepLearning model
}
\description{
Training DeepLearning model, using optimized parameters
}
