% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2o_train.R
\name{train_drf_model}
\alias{train_drf_model}
\title{Training a DRF model, using optimized parameters}
\usage{
train_drf_model(model_name, h2o_data, best_params)
}
\arguments{
\item{model_name}{model_id}

\item{h2o_data}{list of H2OFFrames}

\item{best_params}{optimised parameters}
}
\value{
DRF model
}
\description{
Training a DRF model, using optimized parameters
}
