% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2o_train.R
\name{train_gbm_model}
\alias{train_gbm_model}
\title{Training GBM model, using optimized parameters.}
\usage{
train_gbm_model(best_params, train_frame, val_frame, set, runid)
}
\arguments{
\item{best_params}{optimised parameters}

\item{train_frame}{H2OFrame for training}

\item{val_frame}{H2OFrame for validation}

\item{set}{config object}

\item{runid}{execution ID}
}
\value{
GBM model
}
\description{
Training GBM model, using optimized parameters.
}
