% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2o_train.R
\name{train_glm_model}
\alias{train_glm_model}
\title{Training GLM model, using optimized lambda-value}
\usage{
train_glm_model(model_name, set, h2o_data, best_lambda)
}
\arguments{
\item{model_name}{model_id}

\item{set}{config object}

\item{h2o_data}{list of H2OFrames}

\item{best_lambda}{optimised lambda value}
}
\value{
GLM model
}
\description{
Training GLM model, using optimized lambda-value
}
