% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2o_train.R
\name{train_h2o_model}
\alias{train_h2o_model}
\title{Wrapper function for training final models.}
\usage{
train_h2o_model(h2o_data, set, models)
}
\arguments{
\item{h2o_data}{a list of H2OFrames}

\item{set}{config object}

\item{models}{either a single model object or a list of models}
}
\value{
A list of final model objects
}
\description{
Wrapper function for training final models.
}
