% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2o_train.R
\name{train_isoforest_model}
\alias{train_isoforest_model}
\title{Training a anomaly detection model, using isolationForest}
\usage{
train_isoforest_model(model_name, h2o_data, set)
}
\arguments{
\item{model_name}{model_id}

\item{h2o_data}{list of H2OFFrames}

\item{set}{config object}
}
\value{
isolationForest model
}
\description{
Training a anomaly detection model, using isolationForest
}
