% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{trans_classifyNa}
\alias{trans_classifyNa}
\title{Nominalise features with NA's}
\usage{
trans_classifyNa(df, set)
}
\arguments{
\item{df}{data.frame to transform}

\item{set}{config object}
}
\value{
main object with transformed data.
}
\description{
This module is used, when there is too much NA's in
continuous variables and you can't impute them.
This replaces negative values with \code{'Under0'}, zeros with \code{'0'},
positive values with \code{'Over0'} and NA's with \code{'Missing'} and
factorizes the variable.
}
