% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{trans_delconstant}
\alias{trans_delconstant}
\title{Delete constant features}
\usage{
trans_delconstant(df, set)
}
\arguments{
\item{df}{data.frame to transform}

\item{set}{config object}
}
\value{
main object with transformed data.
}
\description{
Variables that have no variation are dropped.
}
