% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{trans_delequal}
\alias{trans_delequal}
\title{Delete equal features}
\usage{
trans_delequal(df)
}
\arguments{
\item{df}{data.frame to transform}
}
\value{
main object with transformed data.
}
\description{
This function deletes equal columns. Deletes columns with lower indexes.
Returns a list where the first object is a list of vectors that include the
columns that are similar from one loop.
}
