% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{trans_replaceScandAndSpecial}
\alias{trans_replaceScandAndSpecial}
\title{Replace species characters in variables}
\usage{
trans_replaceScandAndSpecial(df, set)
}
\arguments{
\item{df}{data.frame to transform}

\item{set}{config object}
}
\value{
main object with transformed data.
}
\description{
Replace species characters in variables
}
