% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima.R
\name{ts_arima}
\alias{ts_arima}
\title{Function for training ARIMA time series model}
\usage{
  ts_arima(df, set, model_id)
}
\arguments{
  \item{df}{data object}
  
  \item{set}{config object}
  
  \item{model_id}{autogenerated model_id}
}
\value{
  a list containing model details ("arima_model"), its coefficients ("arima_coeff"), evaluation on known data ("arima_eval"), as well as a future forecast ("arima_forecast")
}
\description{
  Function for training ARIMA time series model
}