% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{viz_clusters}
\alias{viz_clusters}
\title{Function for visualizing clustering result.}
\usage{
viz_clusters(D, clust, colors = NULL)
}
\arguments{
\item{D}{distance matrix}

\item{clust}{clustering model}

\item{colors}{color scale}
}
\description{
Function for visualizing clustering result.
}
