% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{viz_silhouette}
\alias{viz_silhouette}
\title{Function for visualizing clustering metrics for a given model.}
\usage{
viz_silhouette(opt, colors = NULL)
}
\arguments{
\item{opt}{clustering model}

\item{colors}{color scale}
}
\description{
Function for visualizing clustering metrics for a given model.
}
