% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{viz_ts}
\alias{viz_ts}
\title{Function for visualizing observed data used in time series model.}
\usage{
viz_ts(h2o_data)
}
\arguments{
\item{h2o_data}{h2o data frame}
}
\description{
Function for visualizing observed data used in time series model.
}
