% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{withWarnings}
\alias{withWarnings}
\title{Catching up warnings-messages}
\usage{
withWarnings(expr)
}
\arguments{
\item{exp}{expression to evaluate}
}
\value{
a list with the evaluation result and warnings
}
\description{
Catching up warnings-messages
}
