% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_data.R
\name{write_csv}
\alias{write_csv}
\title{Write data to file}
\usage{
write_csv(set, df, name, append = TRUE, colnames = FALSE)
}
\arguments{
\item{set}{config object}

\item{df}{data to be written}

\item{name}{name of target file}

\item{append}{wether to extend a file (default = TRUE)}

\item{colnames}{whether to write col.names (default = FALSE)}
}
\description{
Write data to file
}
