% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_data.R
\name{write_db}
\alias{write_db}
\title{Write data to DB}
\usage{
write_db(channel, x, name)
}
\arguments{
\item{channel}{from \code{odbc} object (\code{odbc_metadata})}

\item{x}{data to be written}

\item{name}{name of database table}
}
\description{
Write data to DB
}
