% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{trans_entropy}
\alias{trans_entropy}
\title{Discretisize continuous variables by entropy}
\usage{
trans_entropy(df1, df2, set)
}
\arguments{
\item{df1}{full data}

\item{df2}{combined testing & validation set}

\item{set}{config object}
}
\value{
Discretisize variables. If not possible returns NULL.
Otherwise returns list of discretized data, new variables and cut-points
}
\description{
Discretisize continuous variables by entropy
}
